SIVM-1 Room Database Upgrade

+++ app/src/main/java/com/quigglesproductions/secureimageviewer/models/enhanced/datasource/OnlineFolderDataSource.java
@@ -44,7 +44,7 @@ public class OnlineFolderDataSource implements IFolderDataSource{
     @Override
     public void getFilesFromDataSource(Context context,FolderDataSourceCallback callback, SortType sortType) throws MalformedURLException {
 
-        RequestManager.getInstance().getRequestService().getFolderFiles(folder.getOnlineId(), new RequestManager.RequestResultCallback<ArrayList<EnhancedOnlineFile>, Exception>() {
+        RequestManager.getInstance().getRequestService().getFolderFiles((int)folder.getOnlineId(), new RequestManager.RequestResultCallback<ArrayList<EnhancedOnlineFile>, Exception>() {
             @Override
             public void RequestResultRetrieved(ArrayList<EnhancedOnlineFile> result, Exception exception) {
                 if(exception != null){
